requirements.txt
Flask==1.0.3
Flask-Cors==3.0.8
gunicorn==19.9.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
scikit-learn==0.20.3
scipy==1.2.1
tensorflow==1.13.1
